+++ src/main/java/com/stackmob/sdkapi/SMNotEqual.java
@@ -15,7 +15,25 @@
  */
 package com.stackmob.sdkapi;
 
-/**
- * An update action to take on an object in the StackMob datastore
- */
-public abstract class SMUpdate {}
+public class SMNotEqual extends SMCondition {
+  private final String field;
+  private final SMValue value;
+
+  public SMNotEqual(String field, SMValue value) {
+    this.field = field;
+    this.value = value;
+  }
+
+  public String getField() {
+    return field;
+  }
+
+  public SMValue getValue() {
+    return value;
+  }
+
+  @Override
+  public String toString() {
+    return field + " != " + value.toString();
+  }
+}