+++ RELEASE_NOTES.md
@@ -1,5 +1,25 @@
 # Stackmob Custom Code SDK (Release Notes)
 
+### 0.4.2
+* Added additional query conditions
+  * <, <=, >, >=
+  * Geo queries
+  * Null
+  * Not Equal
+* Order-by
+* Pagination
+* Expand queries
+* Relational queries
+* Get count of objects for a schema
+* Atomic, conditional update one
+
+
+### 0.4.1 (04/30/12)
+* Added additional query functionality
+  * Update by query
+  * IN query
+  * Field filtering
+
 ### 0.4.0 (02/28/12)
 * Added new DataService, replaces deprecated DatastoreService